+++ docs/stories/components/Authenticator.stories.tsx
@@ -1,8 +1,11 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { Authenticator } from "@heathmont/moon-react";
+import {
+  Authenticator as AuthenticatorComponent,
+  FormGroup,
+} from "@heathmont/moon-react";
 import LinksBlock from "../shared/LinksBlock";
 
-type Type = React.ComponentProps<typeof Authenticator>;
+type Type = React.ComponentProps<typeof AuthenticatorComponent>;
 
 const meta: Meta<Type> = {
   title: "Forms & selection controls/Authenticator",
@@ -30,6 +33,20 @@ const meta: Meta<Type> = {
         defaultValue: { summary: "fill" },
       },
     },
+    error: {
+      description: "Defines if Authenticator is in error state",
+      control: "boolean",
+      table: {
+        defaultValue: { summary: "false" },
+      },
+    },
+    disabled: {
+      description: "Defines if Authenticator is disabled",
+      control: "boolean",
+      table: {
+        defaultValue: { summary: "false" },
+      },
+    },
   },
   render: ({ size, variant, ...props }) => {
     const authenticatorProps = {
@@ -37,7 +54,7 @@ const meta: Meta<Type> = {
       ...(size !== "md" && { size }),
       ...(variant !== "fill" && { variant }),
     };
-    return <Authenticator {...authenticatorProps} />;
+    return <AuthenticatorComponent {...authenticatorProps} />;
   },
 };
 
@@ -45,6 +62,29 @@ export default meta;
 
 type Story = StoryObj<Type>;
 
-export const AuthenticatorStory: Story = {
-  args: { size: "md", variant: "fill" },
+export const Authenticator: Story = {
+  args: { size: "md", variant: "fill", error: false, disabled: false },
+};
+
+export const AuthenticatorWithLabelAndHint: Story = {
+  args: { size: "md", variant: "fill", error: false, disabled: false },
+  render: ({ size, variant, ...props }) => {
+    const authenticatorProps = {
+      ...props,
+      ...(size !== "md" && { size }),
+      ...(variant !== "fill" && { variant }),
+    };
+    return (
+      <FormGroup>
+        <FormGroup.Label htmlFor="AuthenticatorWithLabelAndHint">
+          Label
+        </FormGroup.Label>
+        <AuthenticatorComponent
+          {...authenticatorProps}
+          id="AuthenticatorWithLabelAndHint"
+        />
+        <FormGroup.Hint>Hint</FormGroup.Hint>
+      </FormGroup>
+    );
+  },
 };