+++ docs/stories/components/BottomSheet.stories.tsx
@@ -1,8 +1,11 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { BottomSheet, Button } from "@heathmont/moon-react";
+import {
+  BottomSheet as BottomSheetComponent,
+  Button,
+} from "@heathmont/moon-react";
 import LinksBlock from "../shared/LinksBlock";
 
-type Type = React.ComponentProps<typeof BottomSheet>;
+type Type = React.ComponentProps<typeof BottomSheetComponent>;
 
 const meta: Meta<Type> = {
   title: "Containers & layout/Bottom Sheet",
@@ -13,21 +16,31 @@ const meta: Meta<Type> = {
       ),
     },
   },
-  render: ({ ...props }) => {
+  argTypes: {
+    hasHandle: {
+      description: "Shows a handle at the top of the bottom sheet",
+      control: "boolean",
+      table: {
+        defaultValue: { summary: "false" },
+      },
+    },
+  },
+  render: ({ hasHandle, ...props }) => {
     const bottomSheetProps = {
       ...props,
+      hasHandle,
     };
     return (
-      <BottomSheet {...bottomSheetProps}>
-        <BottomSheet.Trigger>
+      <BottomSheetComponent {...bottomSheetProps}>
+        <BottomSheetComponent.Trigger>
           <Button>Open BottomSheet</Button>
-        </BottomSheet.Trigger>
-        <BottomSheet.Content>
+        </BottomSheetComponent.Trigger>
+        <BottomSheetComponent.Content>
           <div className="w-full flex items-center justify-center h-full bg-brand-subtle text-brand overflow-y-auto">
             Content
           </div>
-        </BottomSheet.Content>
-      </BottomSheet>
+        </BottomSheetComponent.Content>
+      </BottomSheetComponent>
     );
   },
 };
@@ -36,10 +49,40 @@ export default meta;
 
 type Story = StoryObj<Type>;
 
-export const BottomSheetStory: Story = {
-  args: {},
+export const BottomSheet: Story = {
+  args: { hasHandle: false },
   play: async ({ canvasElement, userEvent }) => {
     const button = canvasElement.querySelector("button");
     await userEvent.click(button);
+  },
+};
+
+export const BottomSheetWithHeaderAndClose: Story = {
+  args: { hasHandle: false },
+  play: async ({ canvasElement, userEvent }) => {
+    const button = canvasElement.querySelector("button");
+    await userEvent.click(button);
+  },
+  render: ({ hasHandle, ...props }) => {
+    const bottomSheetProps = {
+      ...props,
+      hasHandle,
+    };
+    return (
+      <BottomSheetComponent {...bottomSheetProps}>
+        <BottomSheetComponent.Trigger>
+          <Button>Open BottomSheet</Button>
+        </BottomSheetComponent.Trigger>
+        <BottomSheetComponent.Content>
+          <BottomSheetComponent.Header>
+            Bottom Sheet
+            <BottomSheetComponent.Close />
+          </BottomSheetComponent.Header>
+          <div className="w-full flex items-center justify-center h-full bg-brand-subtle text-brand overflow-y-auto">
+            Content
+          </div>
+        </BottomSheetComponent.Content>
+      </BottomSheetComponent>
+    );
   },
 };